+++ packages/data-generation/package.json
@@ -9,6 +9,10 @@
     ".": {
       "import": "./dist/mjs/index.js",
       "require": "./dist/cjs/index.js"
+    },
+    "./helpers": {
+      "import": "./dist/mjs/helpers/index.js",
+      "require": "./dist/cjs/helpers/index.js"
     }
   },
   "homepage": "https://github.com/beyondessential/tamanu.git#readme",
