--- packages/data-generation/src/index.ts
@@ -1,1 +1,0 @@
-export * from './generateFakeDatabase';
